Post demonstration updates.
Changed search results to json with results of type, and a total property representing total number of results matching the parameters of the request.

Changed 404 response in search no results found to 204.

Changed 409 responses to 400.

Updated date query to replace ':' with '_', in requests, to comply with URL encoding.

Changed add player to session from PUT to POST method